requirements.txt
fastapi==0.103.1
pydantic==2.3.0
python-decouple==3.8
Requests==2.31.0
